Create 10 digit Patient Number
You are a hacker for an insurance company. You would like to know the gender, date of birth, whether patients have medical aid or not, and if they employed or not. You have obtained the 10 digit hospital number of a patient. From this you can obtain all this information. The hospital number is structured as follows:

First Digit: 1 if male, 0 if female
2nd to 8th digit: date of birth i.e. dd-mm-yy
9th digit: 1 if they have medical aid, 0 if they don’t have medical aid
10th digit: 1 if they employed, 0 if they unemployed

Given the hospital number, this program will output the gender, date of birth, whether they have medical aid or not, and if they employed or not.