Create Tampered or Not
When patient data is sent across the hospital, a code is sent to ensure that information is not tampered with. A common test to ensure the message is not tampered with, is to ensure the number is divisible by 3. In order to do this, add the individual digits that make up the number. Then add the digits that make up that number. Keep doing this until a single digit number is obtained. If this value is 3, 6, or 9 then the original number is divisible by 3. Write a program that uses this method to test if a message is tampered with or not.